don't require esModuleInterop downstream
This change updates Node stdlib imports to use the following
ES6 import style:

```
import * as x from 'x';
```

This allows downstream TypeScript projects (direct or transitive)
to use node-agent-base without enabling the esModuleInterop
option. Ideally, it would be nice not to force users to enable
this option if they don't want it.